optimized enrollments

+++ src/course/course.controller.ts
@@ -8,11 +8,13 @@ import {
   Delete,
   Res,
   UseGuards,
+  UnauthorizedException,
 } from '@nestjs/common';
 import { CourseService } from './course.service';
 import { CreateCourseDto } from './dto/create-course.dto';
 import { UpdateCourseDto } from './dto/update-course.dto';
 import { AuthGuard } from 'src/auth/guards/auth.guard';
+import { CurrentUser } from 'src/common/decorators/current-user.decorator';
 
 @Controller('courses')
 export class CourseController {
@@ -20,7 +22,9 @@ export class CourseController {
 
   @UseGuards(AuthGuard)
   @Post()
-  async create(@Body() data: CreateCourseDto) {
+  async create(@Body() data: CreateCourseDto, @CurrentUser() user) {
+    if (user.role === 'STUDENT')
+      throw new UnauthorizedException('Access denied');
     const result = await this.courseService.createCourse(data);
     return {
       message: 'Course created successfully!',
@@ -46,7 +50,18 @@ export class CourseController {
 
   @UseGuards(AuthGuard)
   @Patch(':id')
-  async update(@Param() params: { id: string }, @Body() data: UpdateCourseDto) {
+  async update(
+    @Param() params: { id: string },
+    @Body() data: UpdateCourseDto,
+    @CurrentUser() user,
+  ) {
+    console.log(user);
+    
+    
+    if (user.role === 'STUDENT')
+    
+      throw new UnauthorizedException('Access denied');
+    
     const updated_course = await this.courseService.updateCourse(
       params.id,
       data,
@@ -57,7 +72,9 @@ export class CourseController {
 
   @UseGuards(AuthGuard)
   @Delete(':id')
-  async remove(@Param() params: { id: string }) {
+  async remove(@Param() params: { id: string }, @CurrentUser() user) {
+    if (user.role === 'STUDENT')
+      throw new UnauthorizedException('Access denied');
     const deleted_course = await this.courseService.deleteCourse(params.id);
     return {
       messge: 'course deleted successfuly !',
